Don't read past current TLI during archive recovery (#1189)
If you set the recovery target timeline, the recovery would not
correctly follow the timeline history to reach that timeline, if a
local WAL segment had more WAL records on one of the older
timelines. Fix that by not reading WAL past the switchpoint in the
timeline history.

This case came up with pg_rewind: if you call pg_rewind on a server
that is a direct ancestor of the source, it exits with "no rewind
required" and does nothing. However, if the target had already
streamed WAL past the point of divergence, and just hadn't applied it
yet, when you start it up it will follow the original timeline, not
the one you tried to rewind to. This commit alone doesn't fix that,
because the server still won't know that it's supposed to follow the
other timeline. But this at least fixes the similar case when you have
set the recovery target timeline explicitly.

Discussion: https://www.postgresql.org/message-id/18575-1b9b3b60e2a480de%40postgresql.org

Original patch has been simply modified to correctly apply to our fork:
1) File xlogrecovery.c was renamed to xlog.c
2) Variable flushedUpto was renamed to receivedUpto
3) Callback wait_for_replay_catchup($primary2) was changed to
wait_for_catchup($primary2, 'replay', $primary1->lsn('flush'))