Don't add motions above parameterized sub-plans (#1210)
Finalizing a parameterized subplan in SS_finalize_plan fills the extParam
bitmaps with parameters that affect every node in the plan. But after the plan
is finalized, the planner can add new Motion and Materialize nodes to the
subplan in cdbllize_decorate_subplans_with_motions and fix_subplan_motion. The
bitmaps are not filled for the new nodes. As a result, changing the parameters
stops the subplan from processing the nodes, and it returns a constant result
for each external node, despite the parameter change. Motions do not support
rescans, so adding them at such a late stage of planning is incorrect if the
subplan has external parameters.

Ticket: ADBDEV-6887