Fix assert when using append optimized tables with GIN index (#1221)
Using an AO table with a GIN index could cause an assert in the debug build.
The triggered assert (OffsetNumberIsValid) checks the offset (ip_posid) in the
tuple address (ItemPointerData), indirectly checking the correctness of
GinPostingList decompression. Decompression can be performed, for example,
when inserting data into an AO table with a GIN index or starting a vacuum.
The offset range is checked as for a Heap table. The correct range is 1..8192
with a BLCKSZ of 32768 bytes (see MaxOffsetNumber).
However, in AO tables, the TID offset (ip_posid) is calculated differently,
this led to the assert.

In a Heap table, ip_posid means the ItemIdData number inside the PageHeader
structure. ip_posid for Heap tables is calculated by PageGetMaxOffsetNumber()
macro, for example during insertion. ip_posid can take the value 1..8192 with
BLCKSZ size 32768 bytes.

For AO tables ip_posid is calculated as the value of the lower 16 bits of the
segfile row in AOTupleIdInit() and can take the value 1..32768. What could
cause the assert.

A larger range of ip_posid with AO tuples does not cause problems in the GIN
index, since the full tuple address (ItemPointerData) is used during
processing.

This commit weakens the offset range check (ip_posid) in the assert at the
GinPostingList decompression stage so that AO tables with GIN index can be
processed in the build with asserts.

Changes:
Only the ItemPointerIsValid() check is left;
The appendonly_with_gin_index test has been added to the parallel_schedule;
In the test query, the data volume has been increased and the search condition
has been rewritten so that the GIN index works.

Ticket: ADBDEV-6986