Replace check for Greenplum with check for fork of Greenplum (#1233)
The purpose is to make interaction between Greengage and Greenplum clusters
possible. psql and postgres_fdw checked whether DBMS is Greenplum by analyzing
the output of the version() function. Now they check DBMS for existence of the
'gp_server_version' GUC, which is specific for Greenplum and its forks,
for example Greengage.

Ticket: ADBDEV-7142