Finalize plan after decorate subplans with motions (#1220)
1) Previously, finalization of plans and sub-plans SS_finalize_plan was
performed before decorating sub-plans cdbllize_decorate_subplans_with_motions,
which resulted in newly added nodes not being finalized, including parameter
bitmaps not being set for them. This patch places finalization blocks after the
decorate block, which results in finalization of all plans and sub-plans.

2) When decorating sub-plans cdbllize_decorate_subplans_with_motions, mutation
(including copying) of plans, including init-plans, occurs. Previously,
mutating init-plans did not update the list of init-plans in the root
(root->init_plans), which resulted in desynchronization of this list and the
list of mutated init-plans. This patch updates the init-plan in the root after
its mutation. To do this, it saves subroot in the context.

3) This patch reworks the solution of commit 003bf36 to prohibit adding motions
over parameterized sub-plans. Now, when finalizing a sub-plan, it is checked
whether external parameters are passed through motions. Also, the test outputs
are corrected accordingly.

4) The Postgres planner built an invalid plan in the bfv_catalog test, which,
however, was successfully executed without results with empty tables. Now such
a plan is prohibited at the planning stage. The ORCA planner builds a valid
plan for the same query, which is successfully executed even with data.

Ticket: ADBDEV-7097