Fix aborted database deletion on the entire cluster (#1363)
Commit 034a9fc fixes aborted database deletion on a single instance, and this
patch fixes the same behavior on the entire GPDB cluster. The specified commit
first marked the deleted database as invalid using the heap_inplace_update
function, and only then did the actual database deletion using the irreversible
CatalogTupleDelete function, the call to which was moved further. This is not
enough to do for the entire cluster, because the GPDB first deletes the
database on the segments before deleting it on the coordinator. When resolving
conflicts with commit 45dce8b, this logic was incorrectly left in place, which
led to an error on the entire cluster. This patch moves this logic immediately
before the call to the irreversible CatalogTupleDelete function.

(cherry picked from commit 498fa5d57e13918830657f50cd5c841be9483787)