Fix segfault in utility mode due to uninitialized curSlice (#1446)
27cd38f introduced logic using root->curSlice without checking its presence. 
However, curSlice, which can refer to topSlice, is only initialized on the QD
(GP_ROLE_DISPATCH) in standard_planner(). This caused a segfault in utility
mode during DML CTE execution.

Add a curSlice != NULL check and ensure rootSliceIsWriter is accessed only on
the QD, consistent with the logic around apply_shareinput_xslice().