Fix calculation of block num for page verification in gp_replica_check (#1514)
If the table consists of several segment files, this could lead to an error
when running the gp_replica_check utility. The root cause is that the segment
file number was not taken into account when calculating the checksum of the
page (block) in gp_replica_check. This caused a mismatch between the
checksum in the page header and the calculated checksum of the page. In
order to achieve a checksum match, blockno is passed to PageIsVerified()
taking into account the segment file number.

Ticket: ADBDEV-7418