Fix print KB in explain's output (#1709)
ExplainPropertyInteger() is used to print int64 in explain's output. It requires
int64 variable to correctly convert number to string. At some places double is
used instead of int. Also at some places bytes were not convert to kilobytes,
but were printed with "KB". Patch adds converting double to int64 and bytes to
kilobytes. Variable total_memory_across_slices was removed, because it's value
is not changed and is always zero.

Test was not added, because values of usage memory in the explain's
output are ignored at all tests.