Fix changing locus from SingleQE to OuterQuery (#1163)
cdbpath_create_motion_path() has optimization, which allows not to add extra
motion node for cases, when locus is changed from SingleQE to OuterQuery.
cdbpath_create_motion_path() just changes locus at the first node in the path.
It is correct when path contains just one Motion node. But Outer locus requires
Materialization. The Materialize node must be added before the Motion node.

It is actual for queries with subquery containing both of volatile function and
replicated table. handle_gen_seggen_volatile_path() changes locus to SingleQE
and adds Motion and ProjectionPath node. Next, bring_to_outer_query() changes
locus to OuterQuery and add ProjectionPath with outer quals.

This patch modifies the cdbpath_create_motion_path function for cases when
SingleQE locus is changed to the Outer locus. If the Motion node is the first
node in the path, locus will be changed to Outer and Materialize node will be
added. If there is a Projection path with Motion below, their locus will be
updated, and a Materialize path will be added in between of them. For other
cases, Materialize and Motion paths will just be added on top without modifying
other nodes.

The problem was found when tests from commit
7ef4218221babf046b63f748d3467b54ba5c9ca0 were run. These test are usefull,
because it check correctness of planning queries with volatile functions and
replicated tables. But  another code changes can not be taken because of
difference between 6X and 7X. These tests were added and expanded by adding
execution of queries, which do not contain random().