Fix HAVING pushdown in ORCA (#1589)
According to SQL spec, when HAVING predicate is false, the group should not be
in the output. Specifically, in case of GROUP BY (), output should have no
rows if HAVING predicate is false. However, previously in ORCA the behavior
was different: for example `select count(*) from t group by () having false;`
produced one row containing 0, which is incorrect.

This happened because during expression normalization the HAVING clause
(represented by LogicalSelect above LogicalGbAgg) was pushed through the
aggregation, and so filtering was performed before aggregation, resulting in
an empty group with count(*) of 0. This is valid when there are grouping
columns present, but for GROUP BY () it is not. This patch disables predicate
pushdown in case of no grouping columns.

Other than adding four new ORCA unit tests, this patch also affects several
existing ones:
1. SubqueryInsideScalarIf and CTEWithMergedGroup.mdp: number of plan
alternatives is reduced, since alternatives involving the pushed down
predicate are now forbidden.
2. DqaSubqueryMax: Filtering inside a Join was replaced with JoinFilter, which
shouldn't affect anything.
3. SemiJoin2Select-EnforceSubplan: After commit
652a76f4151f09279d87ea41dab5eae3a503eae2 the plan was changed to something
else without fixing the comment. After this patch, the plan is now back to the
original, matching the comment.

Ticket: ADBDEV-7465