Add missing initialization of bm_next_blkno field (#1723)
In _bitmap_write_bitmapwords_on_page() the wal record (xlrec_perpage structure)
is initialized, which is used to replicate the bitmap index data to the mirror.
Initialization of bm_next_blkno was missed here.
After vacuum, this led to inconsistency and corruption of the bitmap index on
the mirror (the difference between wal record and the bitmap index page data
can be detected in _bitmap_log_bitmapwords()).