Change General to OuterQuery locus in bring_to_outer_query (#1691)
We assume that if we have a correlated subquery for a given path we process
the OuterQuery locus. The patch 605f84e broke this logic - this left the
General locus, even for the correlated subplan (main problem). So we revert it.
Another issue was that the Values ​​Scan node was not treated as OuterQuery
with parameterized qual. This resulted in invalid plans with parameterized
qual in Values ​​Scan if Motion was added on top.
The solution is to allow restrictinfo to be moved to upperrestrictinfo
(see definition) instead of baserestrictinfo for parameterized Values Scan.
Since upperrestrictinfo is not empty, this calls the bring_to_outer_query
handler to distribute the data and prevents to add Motion above the
parameterized node.