Support foreign table column selection as a distribution key (#91)
This patch provides means to mark a subset of columns of a foreign table as the
distribution keys of the table. It allows to eliminate the Redistribute motion
when, for example, the user wants to insert data from a Greengage
hash-distributed table into a foreign table. Before this patch, the insert plan
would contain Redistribute motion, as the foreign table was always assumed as
randomly distributed. With this patch, the foreign table is handled as a
hash-distributed table during planning, allowing the planner to remove
Redistribute motion.

This patch introduces new options for a column definition:
1. 'insert_dist_by_key', which expects a boolean value. It marks the column as
a distribution key.
2. 'insert_dist_by_key_weight', which expects a non-negative integer value. It
sets the weight of the column when the distribution key set contains more than
one column. The overall order of distribution keys is following:
a) all columns that have an explicit weight, sorted by their weights
ascendingly;
b) then all other columns in the order they were set during the table creation.

This feature works only for foreign tables that have 'mpp_execute' set to
'all segments' (either explicitly for the table or derived from SERVER or
FOREIGN DATA WRAPPER options).

Tests are provided in postgres_fdw. A small change in postgres_fdw is introduced
to support the new column options.

Changes from original commit:
- Added a new function GpPolicyFetchExtended instead of modifying original
  GpPolicyFetch. GPDB 7 sets distribution policy for foreign tables in the
  same way as for regular tables, so both INSERTs and SELECTs would be
  affected if original GpPolicyFetch is modified. Instead, make
  GpPolicyFetchExtended and createForeignTablePartitionedPolicy handle both
  cases, and make GpPolicyFetch a wrappar for GpPolicyFetchExtended.
- Add numsegments argument to account for foreign server having different
  number of segments.
- Add SET gp_force_random_redistribution TO on; in tests to disable GPDB 7
  optimization and show Redistribute nodes explicitly for testing purposes.
- Adjust test outputs for GPDB 7.

(cherry picked from commit 6f26515ac0cebd45ed52aa13847342309ca4ba66)