Fix the absence of auxiliary functions due to some isolation2 tests (#78)
Problem description:
Script 'setup.sql' is executed before the isolation2 test suite. It contains
lots of auxiliary functions that are used throughout the tests. But after
running the 'export_distributed_snapshot' test, the auxiliary functions from
'setup.sql' written in Python, like 'pg_ctl()', were no longer available.
That broke any test that attempted to use these functions if it was executed
after the 'export_distributed_snapshot' test.

Root cause:
'export_distributed_snapshot' test dropped cascade the 'plpython3u' language at
its start, and then recreated the language again. It caused all Python
functions created before that moment to be removed.
Also, the issue is relevant for tests 'resgroup/resgroup_dumpinfo' and
'dependency', that also drop the 'plpython3u' language. 

Fix:
Remove the 'drop language cascade' and 'create language' from the
'export_distributed_snapshot' test. The language is created in 'setup.sql' once
for all isolation2 tests. There is no need to re-create it.
Also, do the same for 'resgroup/resgroup_dumpinfo' test. But in the
'dependency' test the drop of the language is done by design, so it can't be
removed. For now we'll keep this test the last one in the schedule, so it
doesn't spoil any other test. A proper disclaimer is added to the schedule file.