Keep skew hashtable in memory for hash joins that require reuse (#95)
When the most common values occuring in the children relations are known, hash
join might enable the skew optimization, placing the most common values into a
separate "skew hashtable", effectively forcing them to be processed in the 0th
batch. This avoids spilling them to disk, which is an important performance
optimization for skewed tables. It is normally okay to free this skew
hashtable after the 0th batch is processed. However when the hash join might
be rescanned, we have to keep it in memory to not lose its contents. This
means that some memory will be freed at the end of the hash join, and not at
the end of 0th batch, but it is acceptable since skew hashtable uses at most
2% of the total hashtable memory.

Ticket: ADBDEV-8454