Make subselect test stable (#90)
Commit cc52dc5 began to use the haveProcessedAnyCalculations field in the
shareinput_mutator_xslice_1 function, including for ORCA. But ORCA at the end
of planning returns a copy of the plan, and copying the object did not copy
this haveProcessedAnyCalculations field. As a result, for ORCA, this field
haveProcessedAnyCalculations came to the shareinput_mutator_xslice_1 function
with an undefined value and caused flaky behavior, including the subselect
test, in which ORCA generated a plan with direct dispatch. This patch adds
setting for ORCA, as well as copying, reading and writing of this field
haveProcessedAnyCalculations, so that for ORCA it now also has the correct
value.

Ticket: ADBDEV-8420