Use current segment state for setup gp-mirror-fast-wait flag (#101)
Commit 5ab4b4e added a new gp-mirror-fast-wait flag to pg_ctl for gpstart. It
is enabled only if the expected state of the segment on the coordinator is
mirror. But the expected state of the segment at the coordinator and the actual
state of the segment may differ, for example, if the mirror is turned off
during a promotion. This patch gets the actual state of the segment, instead of
the expected state on the coordinator, to set the flag.

Ticket: ADBDEV-8661