Stabilize resgroup_views_seg_down isolation test (#120)
Stabilize resgroup_views_seg_down isolation test
Commit 2bd2200 added the test checking query cancellation/termination
behaviour when it's got stuck at checkDispatchResult() at ExecutorEnd()
stage. The changes introduced a new signal sent from FTS when mirror
was promoted after primary's status was detected as down. And at
this signal receiving the hanged qd process wakes up and leaves the
poll() loop in internal_cancel() with further interruptions processing
if such are present.

The corresponding test checks this behaviour in case when FTS promotes
the mirror after the cancellation signal has been sent to a qd proccess.
When test promotes a mirror, it checks the contents of gp_segment_configuration
to ensure that the mirror's status is up. And right after that it checks
that initial hanged query has disappeared from pg_stat_activity. It
assumed that the hanged process could handle the pending interrupts and
canel the query/terminate the backend by the time we query
pg_stat_activity. Mostly this assumption is legit, however, when the test
is running in high concurrency enviroment with high load, the pg_stat_activity
check could show that query is still active. That caused flaky test
output.

In order to stabilize it we wait the session issued the hanged query to continue,
and only then we check the pg_stat_activity.