Fix problem with handling outer references in join clauses (#119)
When the Postgres planner creates a motion for a join inside
cdbpath_motion_for_join(), it looks up the
RestrictInfo.contain_outer_query_references flag for all present
RestrictInfos to determine if a join contains outer references.

The error occurred when original clause was treated as a part of
EquivalenceClass inside distribute_qual_to_rels(), and never
went through the distribute_restrictinfo_to_rels(),
responsible for setting the flag.

To fix the issue, it is enough to set the flag when RestrictInfo
is being created from EquivalenceClass, inside build_implied_join_equality.

Ticket: ADBDEV-7269

Co-authored-by: Denis Garsh <d.garsh@arenadata.io>