Fix inconsistency DROP IF EXISTS command (#138)
When the DROP IF EXISTS command is called, it is unconditionally dispatched to
the segments, even if the object doesn't exist on the coordinator. This can
lead to inconsistencies, for example:

- the first session starts a table drop on the coordinator. Even though the
table doesn't exist, the drop is still dispatched to the segments.

- at the same time, the second session creates the table (on the coordinator
and on the segments).

- the first session (on the segments) already sees the table and therefore
deletes it.

Don't dispatch DROP if object doesn't exist on coordinator. Exclude the
drop_rename test from the parallel group because it contains fault-injectors.
Add a new GUC, gp_dispatch_drop_always, for unconditional DROP dispatch even if
the object isn't present on the coordinator, as this functionality is used by
the gpcheckcat utility, for example, in the orphan_temp_table test.

(cherry picked from commit 3bd88175e9906db0bef54b64ac944cc5eda0f32c)

Changes:

1) The RemoveObjects function no longer has the local variable cell2 and the
condition with stmt->arguments to fill it.

2) The ExecDropStmt function no longer has the OBJECT_EXTTABLE case.

3) The condition on stmt->removeType in the ExecDropStmt function has been
generalized using the shouldDispatchForObject function.

4) The output has been adapted in the drop_rename and direct_dispatch tests.

5) Improved issue reproduction in the drop_rename test.

Ticket: ADBDEV-8908