Fix an array out-of-bounds error in pg_dump
In pg_dump, objects are sorted by priority. Priorities for different types of
objects are stored in dbObjectTypePriority. The values from the
DumpableObjectType enum are used as the array index. It is important that the
number of elements in the array matches the number of elements in the enum.
Previously, the array was missing two elements: DO_EXTPROTOCOL and
DO_BINARY_UPGRADE.

This patch corrects the mismatch by adding missing elements to the
dbObjectTypePriority array. The static assert that was originally added in the
previous commit is returned to pg_dump_sort.