Delete shared tuplestores on abort (#124)
Previously, the creator of the shared tuplestore was responsible for manually
deleting it after it's no longer used. As a result, tuplestores created by
ShareInputScans weren't deleted if an error occured in the middle of a query.
This patch completely reworks the way shared tuplestores are cleaned up.

This patch adds a shared hashtable that stores the sharing state of all shared
tuplestores. The creator of the tuplestore must specify in advance how many
readers/writers will the tuplestore have. Then the tuplestore is automatically
deleted once the last reader/writer closes it, as signified by num_total_left
counter reaching zero.

For error handling, there's an additional global list on each backend,
containing the tuplestores that are currently open. If abort happens in the
middle of a query, all those tuplestores are closed, and deleted if possible.

It's possible that a backend that was supposed to use a tuplestore aborts
before it even opens it. In that case, the num_total_left counter never
reaches zero, and we need to delete the tuplestore prematurely. In this case,
the aborting process is allowed to delete all tuplestores (of the aborting
session) that aren't currently in use, even if future use was expected. For
this, a separate atomic counter num_current is used.

In case when abort happens in a different backend than the one using the
tuplestore, it can't be deleted immediately since it is currently. Instead,
the aborting backend sets an atomic flag in shared memory, and once all the
tuplestore users are done (either while also aborting, or exiting normally),
the tuplestore will be deleted. This way it is not possible for the tuplestore
to not be deleted, since we can guarantee num_current will reach zero once all
backends exit, and that aborting flag will be set before that point.

Since ShareInputScans might be executing inside InitPlans, which are
dispatched as several queries, the EState's lifetime is not enough for these
tuplestores. We instead store them in current transaction's memory context and
resource owner.

The only other place that uses shared tuplestores are functions executing on
initplans. They aren't directly affected by this patch, since the tuplestores
are present on coordinator and would be deleted anyway, but they still get the
new error handling. Since in that case the number of tuplestore users is 2 (1
writer + 1 reader), 2 is chosen as the default value. This case is checked by
function_extensions regress test.

Ticket: ADBDEV-8836