Fix applying Limit for replicated distributions (#105)
The limit path had unnecessary checks for volatility, so in cases of
creating limit over General or SegmentGeneral locuses (replicated
tables or generate_series() for example) limits with just constants
wouldn't pass this checks and generate right path. Replicated locus
was also a subject of this issue and on top of that threw an error in
situations when limit was being used with volatile functions.

There's no need for limit to have volatile function inside of it
to force execution on one segment. Because results of such limit path

->  Limit
    ->  Seq Scan on t_repl

is undetermined in case of execution on each segment (as there is
no guarantees on how seqscan will return it's rows) with constant
expression also.

While applying new logic of limit in correlated subplan of RTE_RESULT
relation, the parameters were being moved through motions, which lead to
an error. This was happening because path creation for trivial case with single
RTE_RESULT relation didn't account situation of having correlated subplan,
like it is accounted for other cases (for example, RTE_RELATION). So this
patch is also fixing that.

---------

Co-authored-by: Viktor Kurilko <v.kurilko@arenadata.io>