Fix orphaned segment files for added and rolled back columns for AOCS tables (#139)
Problem description:
In case we added a new column to a AOCS table in a transaction that was later
rolled back, a non-empty segment file could be left, and vacuum didn't truncate
it.

Root cause:
Vacuum does truncation of AOCS segment files in 2 places:
1. In AOCSSegmentFileTruncateToEOF(). It iterated over 'vpinfo.nEntry', which
doesn't count the added and rolled back attribute, so its segment file was not
processed here.  
2. In AOCSCompaction_DropSegmentFile() function (if compaction is required for
the segment). It iterates over attributes retrieved by
RelationGetNumberOfAttributes(). But, the added and rolled back attribute is not
counted by RelationGetNumberOfAttributes(), so its segment file was not
processed here as well.

Fix:
In AOCSSegmentFileTruncateToEOF() iterate and truncate all segment files for
a given 'segno' until they exist. We stop once we didn't find a file and column
number is above 'vpinfo.nEntry'. We assume that the segment files exist or stop
existing for all columns thereafter - similar logic we already use in
ao_foreach_extent_file().

We do it not in AOCSCompaction_DropSegmentFile(), as it will not be called if
compaction is not required, while the orphaned segment files may still exist in
this case.