Fix QD scan for uncommitted foreign table (#149)
When a foreign table is created, there is a row added to pg_foreign_table,
containing various information including its intended distribution policy.
However, this operation didn't invalidate the relcache, so if a table was
created earlier in the same transaction, the queries wouldn't be able to see
the non-standard execution policy, causing them to be executed in the wrong
location. This was already a known issue for external tables, so cache
invalidation was added by 246f751029e29b74249bfbdf967428a1a5da10c6. Now this
also extends to normal foreign tables, so we better just put cache
invalidation at the end of CreateForeignTable.

The test was added to postgres_fdw for the ease of testing, but the fix
affects all other FDWs as well.

Ticket: ADBDEV-8959