Transition to new cleanup logic for functions on initplan (#153)
The new shared tuplestore cleanup logic introduced by
80b6ac8c6058ebdb68a61a1409b39cd3c3565d6e is now capable for deleting the
tuplestore files when the last reader closes it. We now don't have to make the
writer wait until the reader is done, it can close its copy of the tuplestore
immediately after writing the data. This allows us to remove
postprocess_tuplestore logic in favor of simply calling tuplestore_end
immediately.

However, using this also requires the tuplestore memory to be present during
the abort handler. Executor's MemoryContext and ResourceOwner get freed before
abort handlers are reached, so we need to store all tuplestore data in
transaction-level MemoryContext and ResourceOwner, same as how we did it on
segments already.

Tests are not added since this is a refactor, trying to preserve the current
behavior, not to fix any bug in particular.

Ticket: ADBDEV-8973