Prevent ORCA from losing distribution requests (#157)
Prevent ORCA from losing distribution requests

In the case when the join condition contains AND clauses, hash join performs
redistribution request by every subexpression in the condition.

The problem occurres when several of these subexpressions have the same
right-hand side, making the matching logic inside
CPhysicalHashJoin::PdshashedMatching select the same subexpression
for every request, therefore losing possible distributions.

This patch adds a parameter to the function that specifies the index
of the initial distribution request index. Subexpressions matching
this index are now prioritized in the matching logic.

Also, as a side-effect of the research went into the patch, several
explanatory comments are added to make the code easier to understand.

Ticket: ADBDEV-9011